requirements.txt
feedparser
google-genai